+++ kpermissions/src/test/kotlin/com/fondesa/kpermissions/extension/PermissionsBuilderKtTest.kt
@@ -17,10 +17,12 @@
 package com.fondesa.kpermissions.extension
 
 import android.Manifest
+import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import com.fondesa.kpermissions.builder.PermissionRequestBuilder
 import com.fondesa.kpermissions.request.PermissionRequest
 import com.fondesa.test.createActivity
+import com.fondesa.test.createFragment
 import junit.framework.Assert.assertNotNull
 import org.hamcrest.CoreMatchers.instanceOf
 import org.hamcrest.MatcherAssert.assertThat
@@ -29,27 +31,31 @@ import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 
 /**
- * Tests for CheckPermissionsStatus.kt extensions.
+ * Tests for PermissionsBuilder.kt file.
  */
 @RunWith(RobolectricTestRunner::class)
-class ActivityExtensionsTest {
+class PermissionsBuilderKtTest {
 
-    @Test(expected = IllegalArgumentException::class)
-    fun actPermissionBuilderWithZeroPermissions() {
+    @Test
+    fun `When permissionsBuilder() is invoked with an Activity instance, the PermissionRequest is built successfully`() {
         val activity = createActivity<FragmentActivity>()
-        val builder = activity.permissionsBuilder()
+        val builder = activity.permissionsBuilder(
+            Manifest.permission.SEND_SMS,
+            Manifest.permission.ACCESS_FINE_LOCATION
+        )
 
         assertNotNull(builder)
         assertThat(builder, instanceOf(PermissionRequestBuilder::class.java))
 
-        // This must throw IllegalArgumentException.
-        builder.build()
+        val request = builder.build()
+        assertNotNull(request)
+        assertThat(request, instanceOf(PermissionRequest::class.java))
     }
 
     @Test
-    fun actPermissionBuilderWithSomePermissions() {
-        val activity = createActivity<FragmentActivity>()
-        val builder = activity.permissionsBuilder(
+    fun `When permissionsBuilder() is invoked with a Fragment instance, the PermissionRequest is built successfully`() {
+        val fragment = createFragment<Fragment>()
+        val builder = fragment.permissionsBuilder(
             Manifest.permission.SEND_SMS,
             Manifest.permission.ACCESS_FINE_LOCATION
         )
